refactor: 폴더 구조 개선 및 불필요한 코드 제거
