--- src/app/api/mockAPI.tsx
@@ -1,16 +1,0 @@
-/* eslint-disable no-promise-executor-return */
-
-/**
- * TODO: MSW로 이전하기
- */
-export function fetchMockDeleteSchedule(id: string) {
-  return new Promise<any>((resolve) =>
-    setTimeout(
-      () =>
-        resolve({
-          data: id,
-        }),
-      500
-    )
-  );
-}
